refactor: move trigger API into trigger.internal package
The trigger framework is not yet stable; keep the whole surface internal
so revisions don't break consumers.
