+++ flow-server/src/main/java/com/vaadin/flow/component/trigger/internal/PropertyArgument.java
@@ -13,12 +13,11 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.vaadin.flow.component.trigger;
+package com.vaadin.flow.component.trigger.internal;
 
 import java.util.Objects;
 
 import com.vaadin.flow.component.Component;
-import com.vaadin.flow.component.trigger.internal.ConfigContext;
 import com.vaadin.flow.dom.Element;
 
 /**
@@ -32,6 +31,9 @@ import com.vaadin.flow.dom.Element;
  * <li>{@code Checkbox.checked} →
  * {@code new PropertyArgument<>(checkbox, "checked", Boolean.class)}
  * </ul>
+ *
+ * <p>
+ * For internal use only. May be renamed or removed in a future release.
  *
  * @param <T>
  *            the runtime type of the value produced
