+++ flow-server/src/test/java/com/vaadin/flow/component/trigger/internal/TagComponent.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.vaadin.flow.component.trigger;
+package com.vaadin.flow.component.trigger.internal;
 
 import com.vaadin.flow.component.Component;
 import com.vaadin.flow.component.Tag;
