+++ flow-server/src/test/java/com/vaadin/flow/component/trigger/internal/TriggerSupportTest.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.vaadin.flow.component.trigger;
+package com.vaadin.flow.component.trigger.internal;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -24,7 +24,6 @@ import tools.jackson.databind.JsonNode;
 import tools.jackson.databind.node.ArrayNode;
 import tools.jackson.databind.node.ObjectNode;
 
-import com.vaadin.flow.component.trigger.internal.TriggerSupport;
 import com.vaadin.flow.internal.JacksonUtils;
 
 public class TriggerSupportTest {
